requirements.txt
streamlit
torch
transformers
streamlit-lottie
deep-translator
gtts
Pillow
requests